+++ MealTicket_Web_Handler/Model/GetSharesPlateList.cs
@@ -30,7 +30,7 @@ namespace MealTicket_Web_Handler.Model
         /// <summary>
         /// 涨幅
         /// </summary>
-        public int? RiseRate { get; set; }
+        public long? RiseRate { get; set; }
 
         /// <summary>
         /// 创建时间
